Prevent image distortion/stretching
I noticed that some of the speaker icons looked slightly distorted. I used
the `object-cover` styling for the `object-fit` CSS property. This will maintain
aspect ratio, but parts of the image might be cropped near edges to achieve this.
When I applied this change the image looked less distorted, but the faces were
all still visible.

I also applied the `object-contain` property to the header icon so that it will
also maintain its aspect ratio. I had noticed on small screens that it would get
slightly distorted in the horizontal direction. I believe this change should fix
that.