Pull text colors to parent level and support dark mode
This pulls text colors out so that most text elements are defaulted
to what is specified in root.tsx. I don't think we would leave the
theme mode button in the top right for switching modes (I'm guessing
we would just use the system default). In the future we can refine
the background and text color look, but for now we can plow forward
without making too big of a mess with varying text styles everywhere.