Refactor the header and footer to server side
In order to have a search bar in the header and breadcrumb navigation,
we first need to move the header to the server side (so we can load
data from the "remix" node process). It made sense to move everything
to the server side including the footer and margin styling. This was
done through an SiPage component that every route (page) should be
wrapped with (unless of course we have need of a route/page that we
want custom styling and no header/footer, this of course, will only
be possible with the changes in this commit.