Add simple breadcrumb navigation
This adds simple breadcrumb navigation. We will likely need to
continue to upkeep this function as we add bible routes or other
routes. I decided to pass in optional props, like the sermon type
opposed to fetching the sermon when inferring the nav crumbs. I
was hitting a CORS security issue when trying to fetch the sermon
from the Breadcrumb component because it is not a "route". I think
the best pattern for remix is to have each "route" be a page, and
for each page to have a loader that loads the html on the server
side, however, any sub-component within that react node should be
passed the props opposed to doing sub-loading or sub-fetching.