Fix two playlist mode bugs
When clicking an item in the list, it did not start playing. I changed the
behavior to make it start playing.

The second bug fix is that now the window can be backgrounded (minimized)
and the media player will continue to play through the list. Previously it
would play through the current item and then not start playing the next item
until the window was foregrounded.