Tried to share heading styles between the blog and SiSection(s)
This also formats the bible verse page for desktop with the exception of
navigation arrows and a bug when linking from the commentaries. This
commit also includes a bunch of other formating related changes... hopefully
most for the better? I was mostly concerned with sharing the headings and
might have lost track of what some sections were supposed to look like.